Hex Highlighting
The _Hex class now has a more general highlighting system that can be operated using the `highlightHexes` method of the `HexGrid` class.

The movement range highlighting has been overhauled to use the new highlighting system.

There is now a unit-selection-system that is handled by the UnitManager. This system is operated by the hex-selection-system (since the user can only select hexes).

The direction members of the _Hex class are now actually pointing in the correct directions!

There is a new function to get a hex which is intended to be use solely in the AGeIDE console

Many casts to int have be replaced with `round` to ensure that 3.99999 would be correctly converted to 4