End of first stage (and Combat System Test)
Adds a (for this point in the development far too complex) combat system but without any weapons....
This system is only a first test to see how to implement a combat system.

This commit marks the end of the first stage of this project. Now that the basis exists that could be quickly developed into a playable proto (meaning very rudimentary) game it is time to cut this basis into multiple files so that it can be turned into a proper large-scale game.