Start of Stage 2
Renaming Main.py to _Main_old.py marks the beginning of Stage 2.
In Stage 2 the project will be split into multiple files and folders and the base infrastructure of the game will be established. The goal is to establish all base classes and base systems.
Stage 3 can then use this basis to build an actual game.
Stage 2 is mostly about organisation and less about making game-features.