Start of Division
The hexagon things, the Unit-class, the getter, and the window are now in their own files. There is also a new widget file.

There are now shields and weapons effects and a very simple gui to display the stats of ships.

Ship models are now automatically scaled and centred.

Main_temp.py is currently the main file. the new Main.py is empty for the moment. It will get used once a strategic map is implemented.

There are some other minor changes and improvements but they are all too minor and there are especially too many to list them all (and I have forgotten most of them...)