Basic Classes
All of the code form `Main_temp.py` is now in proper modules.
This commit furthermore adds the basic classes for ships which replace the `Unit` class. This marks the end of the first prototype stage (and breaks all of the temporary test-features).
`Main_temp.py` is now outdated and is probably broken. `Main.py` now assumes the role as the main file and `Main_temp.py` will be removed in the very near future.
The next step will be to take all of the functions of the temporary `Unit` class and implement them in the proper classes. (Most methods are already copied to the fleet-class and await a complete overhaul.)