It Works Again
The overhaul is now at a point where the Main.py works again.
flotillas, ships, and  the hull module now work. The hull module will probably get removed down the line as ships should be infinitely expandable and the stats should derive from the modules that make up the ship. But for now the hull module is a nice way to test the system.

Most importantly there is now a Ships folder and a ShipModules folder which can now be filled with content to build up a first MVP for the battle system.