Stuff is happening!
I have reworked how most thing work.
It is a bit much to write into this commit message...
We now have Modules, a GUI, meaningful Turns, and many things have changed.
I am currently working on the GUI It is currently a WIP and I actually forgot to even push the previous iteration of the GUI.
I think I am currently in a sort of greenfield-project-rush. Let's hope that progress keeps happening at this pace!