Construction Window Version 1
There is now a dedicated construction module to design and build you own ships. The current version does work but many features are still missing, the UI is very unfinished, and there are no checks to prevent user error! For example: if you add a second hull to the ship the first hull will get removed from the ship but not from the UI. Removing it afterwards may or may not work. The Window also stores references to all modules and to the ship which may lead to weird behaviour when a ship gets destroyed. And the file that contains the construction window's code is reloaded every time the window is opened which might cause problems (but is super useful for development). And I have not tested if the destruction of the construction module is handled as I intended it. And there are probably even more problems that I have not even considered.
But it is at least a good starting point!