Harvestable Resources
This commit adds harvestable resources in the form of resource rich asteroids on the campaign map as well as modules to harvest and process these resources.
There are also many other minor changes, too numerous to list them all here.
There was about a year in which I didn't work on the game so the development of this update had a huge gap in the middle.
This means that I can't quite remember what needs to be tested.
And since I am currently not in the mood to fully playtest the current version of the game but instead want to work on the next one, there may be many bugs in this version, so beware the bugs.
I don't even know this the save and load functionality still works but I doubt anyone will play before I have done a proper playtest and fixed everything.