Pure Shader Skybox
The generation of the old skybox was too slow for my taste and ate up unnecessary memory. This commit implements a new skybox that is fully implemented as a shader. As this might be a problem on older graphics cards there is still an option to switch back to the old skybox generator.