Show Dev Tool Tabs Button
Hides the dev tool tabs (console, inspect, etc.) and adds a button to the debug tab that shows the dev tool tabs. This makes the tabs easier to navigate and read.
Fixes a bug that would cause the jump modules to break the ship designer.
Removes neutral objects from the ship selector, making the "Copy existing ship" menu in the ship designer more usable.