New WIP star shader function (unused for now)
I started work on a new WIP shader function to draw the stars. The old one is currently still used, but the new one should produce much better results, once it's finished, and should also be usable to draw the larger stars.
Also added a Version of the original star function with some minor fixes that still need a bit of testing and tweaking and will probably get deleted soon but I wanted the fixes in the version history in case I need it in the future.