Zoom To Cursor And Cam Fix
Implements zoom-to-cursor and adds an option for it (default on).
Fixes the a problem that would occur then binding the cursor to the window for camera control when clicking MMB while the cursor is moving fast.
The option to bind the cursor while controlling the camera is now on by default on all platforms except for windows as it apparently does not work there (according to the official panda3d docs).