Fix weird exec bug
So, those exec statements have always used the globals from the module. But all of the sudden the ShipBase.copy method no longer works because ShipBase is already in the globals but it is expected that ShipBase is a module that is imported and ShipBase.ShipBase would need to be accessed. Nothing has changed so I don't understand why it suddenly broke after 11 months (I can't believe that I never tested this feature in the past 11 months so something must have broken it very recently. I think I have tested it just 4 days ago.).
Anyways, since all of the exec statements don't actually need globals and should use the same dictionary for globals and locals anyways (for imports being accessible in function) that is the way they work now (they get the same empty dict for both locals and globals). This should have been the way they should have worked from the start and now I am a little confused that they were working previously in the first place...