Flotilla splitting fix and more
Fixed a bug when splitting a fleet into flotillas that would not put all ships into a flotilla.
Added more error handling to flotilla splitting while I was at it. Probably superfluous but better save than sorry.
Made resources more floating point error resistant.
Doubled resources in asteroids.
The construction window now shows the cost of modules when no changes are to be applied
The construction window now shows the difference between available resources and ship costs to make planning easier
Increased the default strength of enemy fleets in the menu as they seemed a bit to weak.
Added a tooltip to the End Turn button that shows the hotkeys.