Faster Static Skybox
Make the static skybox generation faster by not saving it to disk and then loading it for display (took me way too long to figure out how to skip the save/load step but it finally works now (though it is still slightly slower than expected...)).
Made static skybox the default as it is way less demanding on the graphics card.
Increased the default static skybox resolution as it is now fast enough.
Make sure all random numbers in the skybox generation use the seed (especially the colourmixing for the bright stars was missing the seed).
Make sure the window is properly focused on windows (might need further tweaking)