Experimental CPU Nebulae for Shader
This commit adds some experimental code for generating the nebulae on the CPU for the pure shader skybox. This is still only a very early prototype but it drastically reduces the GPU load for the pure shader skybox and could be a promising solution for lower end systems.
But the current version looks very bad and still needs a lot of work...