Ships Scale According to Mass
Ships in fleets now scale according to the mass of the individual ships in the fleet.
Models of single ships  (fleet size == 1) now scale with respect to the largest ship on the current map to ensure that a small fighter and a large carrier that are alone and next to one another actually look proportional to another.
(Note: I originally wrote this code almost a year ago but only came around to finalizing it now. glad to finally have this committed now and longer having to exclude it from commits.)