Background Objects and Blocking Tiles
Introduces the concept of Background Objects, which are basically all objects that populate the map which are not normal fleets, like asteroids and nebulae.
There can be multiple Background Objects per tile, but only one normal fleet (Background Objects are still organized as fleets).
Fleets that only contain Ships with the flag Background Object are considered background fleets and are stored in a tiles Contents list instead of the Fleet reference.
Also introduces the concept of Blocking for fleets. Ships can be Partially Blocking, Completely Blocking, both, or neither.
A tile with something partially blocking can be moved onto by things that are not blocking but tiles with something that is completely blocking can not be moved onto by anything. Though tiles with a non-background fleet can not be moved onto by another non-background fleet.
These 3 new flags are already implemented in the save-system and resource harvesting system, but they still need more testing.
While a tiles Content (i.e. Background Objects) are now displayed in the economic view if they contain resources, I still need to implement the normal non-economic interface.
While this commit should not break any save file compatibility, existing asteroids will not automatically be converted to Background Objects.