Harvestable Debris
Instead of being dumped on the tile, salvage is now stored in a debris Background Object.
NOTE: The Debris currently has no model an will be displayed as a big vision blocking cube that is very annoying. I will add a model to debris soon.
Changes the node a fleet's node is parented to from the node of the active scene to the node of the appropriate (Battle/Campaign) scene.
This commit contains some other fixes and improvements, especially regarding fleet movement