Fix Debug Tab Button Bug
When clicking the button to show the debug tabs the button now gets properly removed from the UI (well it got removed previously but that meant it would just flout uncontrolled in the UI but mostly behind another button so I only noticed now. Now the Button is also hidden effectively remove it. I probably should properly delete the button but it is merely a debug button so who cares...)