More Ship Model Variety
I got bored of all these blocky ships so I tried to convince this newfangled ChatGPT thingy to write me code to generate more shapes and after quite some tweaking and fixing I got some of them to actually work and partially even look somewhat good. Not quite happy with all of them, yet, but it's at least a nice starting point.
Currently each ship gets assigned (not explicitly but incidentally due to how I have implemented the seeds) a shape that is then used for all models.
I have already implemented an interface to be able to assign each ship and each team a specific style, but that interface is currently unused.