Team Styles
Rudimentary implementation of team styles by assigning The Star Nomads the standard block, The Buccaneers the circular shapes, and The Star Raiders the more complex block shapes. I will need to overhaul this in the future but it at least is a temporary solution that makes use of the new styles to give the teams more of an identity and avoids the not-so-pretty randomness from before.