Minor Fixes and Improvements
Reduces occurrences of 'hex.fleet().addShip'.
When adding ships to a Background fleet the fleet now checks if it should become a foreground-fleet.
Adds a try-catch to a problematic section of the UI-code.
Adds a new Debug button to destroy selected fleets.
Move the code for removal of debris from the debris ship class to the debris module class.