Construction Window Module Categories
The modules in the module list of the construction window are now split up into categories by using a QTreeWidget.
The modules in the list also have a Tooltip now which shows the module name (useful if name is too long to fit in the list) and the cost of the module (useful in live-editing mode).
Added a new module class _FundamentalModule which is now the parent class for Hull, Engine, Thruster, and Sensor Modules to make them easier identifiable using issubclass.