Construction Window Cost Colourisation
All applicable resource costs in the construction window are now red when there are not sufficient resources available and green if a change would yield resources (e.g. when life-editing a ship and decreasing the stats of a module, making it cheaper).
Overhauled _ResourceDict.text() to provide a rich-text formatting option as well as colour highlighting based on comparisons to a given other _ResourceDict to show if enough resources are available. (Colours for highlighting are taken from the new "Special Text Colours" entry for AGeLib Themes.)
Overhauled creation of ConstructionWindow to allow for providing the ConstructionModule on initialisation to ensure that the module list can check if adding a given module would be affordable when the module list is first populated.