AI Movement Improvement
When the AI wants to move near a hex that is far away the pathfinding now takes modified movement cost into consideration. For this I have modified findPath in HexBase.py to be able to only return the part of a path that can be afforded using a given amount of movement points.