Resource Transfer Selection Overhaul
Selecting items in the resource transfer widget with left-click now cycles them internally (i.e. new selection replaces oldest selection) to be more intuitive. Right-click now always replaces the older selection but marks it as the older selection thus preserving the newest left-click selection which makes collecting resources in one item easier.
The Transfer Window now opens properly (reasonable size in the middle of the screen instead of being tiny in one corner of the screen).
I ahve also tried to debug the transfer window closing when the palette changes but it seems to have nothing to do with the code of the transfer window. Might be a Qt bug...