Resource Transfer Prohibit Nested Selecction
Selecting an item inside/outside an already selected item in the resource transfer window now removes the selection of the outer/inner item and only selects the inner/outer item. This not only makes the highlighting work (as the selection of the inner item would be indistinguishable from the outer one) but crucially prevets the case where resources would be transferred from an inner container to its outer manager who would just put the resources back into the container (or the other way round).